ci: enforce UTF-8 and add CI test fixtures
Force UTF-8 for stdout in .github/scripts/inject_ci_flag.py to avoid crashes on legacy Windows terminals; streamline the custom_features injection and harden error printing. Add tests/ci/conftest.py which provides a repo_tmp fixture that yields a project-root-relative POSIX temp dir for WSL compatibility. Update tests/ci/test_ci_flag_injection.py to call the injector via an absolute path, set PYTHONIOENCODING, use UTF-8 output capture, and improve failure assertions. Remove the duplicated repo_tmp fixture from test_salt_injection.py.
